Backend/requirements.txt
flask
flask_sqlalchemy
flask_jwt_extended
requests
pytesseract
Pillow
werkzeug
